third_party/WebKit/LayoutTests/svg/animations/animate-path-animation-Ll-Vv-Hh-expected.txt

SVG 1.1 dynamic animation tests
Test path animation where coordinate modes of start and end differ. You should see PASS messages

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS path.getAttribute('d') is "M -30 -30 L 30 0 V 30 H 0 Z"
PASS path.getAttribute('d') is "M -15 -15 L 15 0 V 15 H 0 Z"
PASS path.getAttribute('d') is "M 15 15 l -30 -15 v -15 h 15 Z"
PASS path.getAttribute('d') is "M 29.985 29.985 l -59.97 -29.985 v -29.985 h 29.985 Z"
PASS path.getAttribute('d') is "M -30 -30 L 30 0 V 30 H 0 Z"
PASS successfullyParsed is true

TEST COMPLETE
